Fix devise and better_errors versions in Gemfile
Add specific version limits to devise and better_errors gems, so that
they do not upgrade to breaking version numbers, and so that new
installations do not run into errors.  The latest version of
better_errors wants Ruby 2 (while we are using 1.9), and devise v3+
requires changes to our configuration to avoid errors on startup,
and is untested with our application.

A bundle update on an old installation will trigger the errors mentioned
above.