Pin rake gem
The `rake` gem has had a new release that obsoletes some syntax that's
used in our application. The error below is encountered when running
`rake assets:precompile`.

See
http://stackoverflow.com/questions/35893584/nomethoderror-undefined-method-last-comment-after-upgrading-to-rake-11

The solution is for us to run version 11.3.x of `rake`, which we've
known to be good in production.  Thus, Gemfile is being modified in this
commit to pin it.

The error message:
NoMethodError: undefined method `last_comment' for <Rake::Application:0x00000001cce808>
/home/dpla/frontend-git/Rakefile:8:in `<top (required)>'