Improve response times of various facets
Improve response times of the Timeline and other places where very large
facets have been queried, processed, and rendered in the past.

* Reduce facet sizes
* Create querystrings more efficiently in self-referential hyperlinks in
  certain places that have tight `each' loops, where `link_to' is too
  slow.

Note that the huge facets of 2000 elements were never used fully,
because the most that was drawn into a given modal window on an HTML page
was 10 "pages" of 30 elements. Only 300 facets were really being used out
of the 2000 that Elasticsearch was being asked to assemble.