Add self-hosted Universal Viewer
* Change script and div tags to use our locally-hosted installation.
* Add our own configuration JSON file for the viewer.

Note that this change requires that we install the Univeral Viewer
Javascript by whatever means. In our case at the DPLA we install it
through our `automation` configuration management and deployment system.

The configuration file is managed outside of the viewer installation so
that different web applications that embed it can theoretically have
their own settings.